Bugfix: escape only email tags, even if attribute values  contain an @ character
Email clients use an email-like notation to generate content-IDs for embedded images. The content-ID therefore contains an at character (@).  Unfortunately, the html_sanitize mistakes tags with such attributes for "email tags" (meant for mike <mike@example.com>) and causes the html code to be displayed literally.

This patch requires that the tag contains no more double quotes after the @, separating the usage of @ in an attribute value from the case with the email tags.